+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+- Remove support for Python 2, Django 1.11,2.0,2.1, DRF 3.7,3.8,3.9
+- Add support for Django 3.0, DRF 3.10,3.11
 
 ## 0.1.1
 - Fix passing arguments to etag_func and last_modified_func